requirements.txt
discord.py
google-cloud-secret-manager
py-cord==2.0.0b5